pyproject.toml
[project]
name = "prueba-tecnica-ceiba-software"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "db-dtypes>=1.4.4",
    "google-cloud-bigquery>=3.41.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
]
